+++ src/components/docs/docs-content-layout.svelte
@@ -4,7 +4,11 @@
   import FeedbackWidget from "./feedback-widget.svelte";
   import EditInGitpod from "../../components/docs/edit-in-gitpod.svelte";
   import docsCurrentSectionStore from "../../stores/docs-current-section";
+  import { page } from "$app/stores";
+  import { getMenuContext } from "../../routes/docs/menu";
+  import PrevNext from "./prev-next.svelte";
 
+  const menuCtx = getMenuContext($page.path);
   export let section;
   export let title;
 
@@ -23,4 +27,5 @@
   <EditInGitpod />
   <slot />
   <FeedbackWidget />
+  <PrevNext {menuCtx} />
 </div>
